Fix some style issues and factor out old address lowering provider
